refactor(program-doc): split into types/style/ui per new conventions
