+++ src/features/program-doc/index.ts
@@ -1,1 +1,2 @@
 export { default as DOCProgram } from "./DOCProgram";
+export type { ProjectData, ProjectResult, ProjectStack } from "./DOCProgram.types";
